+++ src/aten/sycl/UnaryKernels.cpp
@@ -85,32 +85,6 @@ struct ExpFunctor {
 };
 
 template <typename scalar_t>
-struct SgnFunctor {
-  scalar_t operator()(scalar_t a) const {
-    using opmath_t = at::opmath_type<scalar_t>;
-    if (a == scalar_t(0, 0)) {
-      return scalar_t(0, 0);
-    } else {
-      return a / std::abs(static_cast<opmath_t>(a));
-    }
-  }
-};
-
-template <typename scalar_t>
-struct SignFunctor {
-  scalar_t operator()(scalar_t a) const {
-    return c10::signum(a);
-  }
-};
-
-template <>
-struct SignFunctor<bool> {
-  bool operator()(bool a) const {
-    return a;
-  }
-};
-
-template <typename scalar_t>
 inline scalar_t reciprocal_wrapper(scalar_t a) {
   return static_cast<scalar_t>(1) / a;
 }
@@ -316,25 +290,6 @@ void exp_kernel(TensorIteratorBase& iter) {
         [&]() {
           auto caller = ExpFunctor<scalar_t>();
           gpu_kernel(iter, caller);
-        });
-  }
-}
-
-void sgn_kernel(TensorIteratorBase& iter) {
-  AT_DISPATCH_COMPLEX_TYPES_AND(kComplexHalf, iter.dtype(), "sgn_xpu", [&] {
-    using opmath_t = at::opmath_type<scalar_t>;
-    gpu_kernel(iter, SgnFunctor<opmath_t>());
-  });
-}
-
-void sign_kernel(TensorIteratorBase& iter) {
-  auto dtype = iter.dtype();
-  if (dtype == ScalarType::Bool) {
-    gpu_kernel(iter, SignFunctor<bool>());
-  } else {
-    AT_DISPATCH_ALL_TYPES_AND2(
-        ScalarType::Half, ScalarType::BFloat16, dtype, "sign_xpu", [&] {
-          gpu_kernel(iter, SignFunctor<scalar_t>());
         });
   }
 }
