+++ src/aten/sycl/UnaryKernels.h
@@ -24,8 +24,4 @@ void bitwise_not_kernel(TensorIteratorBase& iter);
 
 void exp_kernel(TensorIteratorBase& iter);
 
-void sgn_kernel(TensorIteratorBase& iter);
-
-void sign_kernel(TensorIteratorBase& iter);
-
 } // namespace at::native::xpu
